% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{cluster}
\alias{cluster}
\title{Generate template for clsutering, counting and plotting colors}
\usage{
cluster(from, to, shortname = paste0(from, "_", to))
}
\description{
Generate template for clsutering, counting and plotting colors
}
