% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{count_colors}
\alias{count_colors}
\title{Count colors based on proximity to centers}
\usage{
count_colors(colors, centers, colorspace)
}
\arguments{
\item{colors}{colors to be tabulated as a \link[colorspace:color-class]{colorspace::color} object,
e.g. \link[colorspace:RGB]{colorspace::RGB}.}

\item{centers}{colors to be used for binning as a \link[colorspace:color-class]{colorspace::color} object,
e.g. \link[colorspace:RGB]{colorspace::RGB}.}

\item{colorspace}{string describing the colorspace in which distances should
be calculated, e.g. \code{"RGB"}, \code{"LAB"}, \code{"LUV"}.}
}
\value{
tibble with as many rows as \code{centers} with the columns \code{freq} and
\code{hex}, as well as columns corresponding to the dimensions of \code{colorspace}
}
\description{
Colors in \code{colors} are binned to the closest colors in \code{centers} based
on Euclidean distance in the color space provided in \code{colorspace}.  These
are then tabulated to give the number of pixels of each color in \code{centers}.
}
\examples{
example_jpg <- system.file("extdata", "example.JPG", package = "routes")
read_pixels(example_jpg) \%>\% 
  colorspace::RGB() \%>\%
  count_colors(centers = colorspace::RGB(c(1, 0, 0), c(0, 1, 0),  c(0, 0, 1)),
    colorspace = "RGB") 
}
