% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{display_jpeg}
\alias{display_jpeg}
\title{Display JPEG in graphics device}
\usage{
display_jpeg(file)
}
\arguments{
\item{file}{path to JPEG file}
}
\value{
Invisibily the native JPEG object.
}
\description{
A way to quickly check the contents of a JPEG file on disk.
}
\examples{
example_jpg <- system.file("extdata", "example.JPG", package = "routes")
display_jpeg(example_jpg)
}
