% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{equal_margins}
\alias{equal_margins}
\title{Set equal ggplot2 margins}
\usage{
equal_margins(margin = ggplot2::unit(0.5, "inches"))
}
\arguments{
\item{margin}{\code{\link[ggplot2:unit]{ggplot2::unit()}} object with required margin size.}
}
\value{
list appropriate for adding directly to ggplot2 object
}
\description{
Sets theme and scales so that the blank area around the plotting area is the
same width on all four sides.
}
\examples{
library(ggplot2)
rect <- data.frame(x = 0, y = 0)
p <- rect \%>\% 
  ggplot(aes(xmin = x, xmax = x + 1, ymin = y, ymax = y + 1)) +
    geom_rect(fill = "black")
p     
p + equal_margins()
}
