% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\name{get_img}
\alias{get_img}
\title{Get Street View image from Google Street View API}
\usage{
get_img(lat, lon, dir, api_key = maps_api_key(), size = "100x100",
  suffix = "", ...)
}
\arguments{
\item{lat}{latitude}

\item{lon}{longitude}

\item{dir}{path to directory image should be downloaded to}

\item{api_key}{Character string of your API key.  By default, looks for
\code{GOOGLE_API_KEY} environment variable, see \code{\link[=maps_api_key]{maps_api_key()}}.}

\item{size}{dimensions in pixels of file to download as a string, e.g. "100x100"}

\item{suffix}{string to be appended to the end of the file name}

\item{...}{other parameters passed along to the Google Street View Static API,
e.g. heading, pitch.}
}
\value{
file path of downloaded file, invisibly.
}
\description{
Download Street View image at given latitude, longitude location.
}
\details{
See the \href{https://developers.google.com/maps/documentation/streetview/intro#url_parameters}{Google Street View Static API Docs}
for more info on available parameters.
}
\examples{
\dontrun{
img <- get_img(lat = 37.2341762, lon = -112.8726299, dir = tempdir())
display_jpeg(img)
}
}
