% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\name{get_route}
\alias{get_route}
\title{Get route from Google Directions API}
\usage{
get_route(from, to, api_key = maps_api_key(), ...)
}
\arguments{
\item{from}{Character string for start of route: "The address, textual latitude/longitude
value, or place ID from which you wish to calculate directions".}

\item{to}{Character string for end of route}

\item{api_key}{Character string of your API key.  By default, looks for
\code{GOOGLE_API_KEY} environment variable, see \code{\link[=maps_api_key]{maps_api_key()}}.}

\item{...}{Other parameters passed to the
\href{https://developers.google.com/maps/documentation/directions/intro}{Directions API}}
}
\value{
tibble with \code{lat}, \code{lon} and \code{order} columns for each step in the
route.
}
\description{
Coordinates of a route from two locations from the Google Directions API.
}
\details{
The coordinates match instructions in the route and they will not
be equispaced along the route.
}
\examples{
\dontrun{
get_route("Corvallis, OR", "Portland, OR")
get_route("37.2341762,-112.8726299", "37.2174425,-112.973322")
}
}
