% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\name{maps_api_key}
\alias{maps_api_key}
\title{Get Google Maps API Key}
\usage{
maps_api_key()
}
\description{
Looks for API key in environment variable \code{"GOOGLE_API_KEY"}
}
\details{
routes assumes you have a Google API key stored in the environment variable
\code{GOOGLE_API_KEY}.  First, get an \href{https://developers.google.com/maps/documentation/streetview/get-api-key}{API key from google},
then run:\preformatted{usethis::edit_r_environ()
}

to open your \code{.Renviron} file and add a line of the form:\preformatted{GOOGLE_API_KEY="XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
}
}
