% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{pal}
\alias{pal}
\title{Plot a palette of colors}
\usage{
pal(col, border = "light gray", ...)
}
\arguments{
\item{col}{character vector of hex color codes}

\item{border}{color for border around swatches}

\item{...}{additional arguments passed to \code{plot()}}
}
\description{
Plot colors, \code{col}, as a single row of swatches.
}
\examples{
pal(c("#8FA5DB", "#AB9F83", "#95B9E5"))
}
