% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{read_pixels}
\alias{read_pixels}
\title{Read pixels from JPEG}
\usage{
read_pixels(file)
}
\arguments{
\item{file}{path to JPEG file}
}
\value{
matrix with three columns, R, G and B, and as many rows as there are
pixels in the JPEG.
}
\description{
Reads pixels from a JPEG file into a matrix with RGB columns
}
\examples{
example_jpg <- system.file("extdata", "example.JPG", package = "routes")
read_pixels(example_jpg)
}
