% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{route}
\alias{route}
\title{Generate template for getting coordinates for route}
\usage{
route(from, to, shortname = paste0(from, "_", to))
}
\description{
Generate template for getting coordinates for route
}
