% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{sample_pixels}
\alias{sample_pixels}
\title{Sample pixels from a JPEG file}
\usage{
sample_pixels(file, sample_size = 50)
}
\arguments{
\item{file}{path to JPEG file}

\item{sample_size}{Number of pixels to sample}
}
\value{
tibble with \code{sample_size} rows, and columns R, G and B
}
\description{
Takes a simple random sample of \code{sample_size} pixels from a JPEG file.
}
\examples{
example_jpg <- system.file("extdata", "example.JPG", package = "routes")
set.seed(4891)
sample_pixels(example_jpg, 10)
sample_pixels(example_jpg, 10) \%>\% 
  with(colorspace::RGB(R, G, B)) \%>\% 
  colorspace::hex() \%>\% 
  pal()
}
