% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup directory and file for generating a route}
\usage{
setup(from, to, shortname = paste0(from, "_", to))
}
\arguments{
\item{from}{address, or latitude and longitude of the start of the route.}

\item{to}{address, or latitude and longitude of the end of the route.}

\item{shortname}{string suitable for use as directory name and in file names.}
}
\value{
Called for it's side effects only. Creates a directory called
\code{shortname}, creates a file in this directory
called \code{00-setup.R} populated with a template.
}
\description{
The first step in creating a new route landscape.
}
