% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{streetview}
\alias{streetview}
\title{Generate template for getting streetview images}
\usage{
streetview(from, to, shortname = paste0(from, "_", to))
}
\description{
Generate template for getting streetview images
}
