% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{streetview}
\alias{streetview}
\title{Generate template for getting streetview images}
\usage{
streetview(from, to, shortname = paste0(from, "_", to))
}
\arguments{
\item{from}{address, or latitude and longitude of the start of the route.}

\item{to}{address, or latitude and longitude of the end of the route.}

\item{shortname}{string suitable for use as directory name and in file names.}
}
\description{
Generate template for getting streetview images
}
